Restructure public API
