+++ src/lib.rs
@@ -1,18 +1,15 @@
-pub use async_trait::async_trait;
-
 mod bytes_ext;
 mod error;
 mod handle;
-pub mod packet;
 mod parse;
 mod read_req;
-mod server;
 mod server_builder;
 mod tests;
 #[cfg(feature = "unstable")]
 mod write_req;
 
+pub mod packet;
+pub mod server;
+
 pub use crate::error::*;
-pub use crate::handle::*;
-pub use crate::server::*;
-pub use crate::server_builder::*;
+pub use async_trait::async_trait;
