+++ src/server.rs
@@ -13,11 +13,14 @@ use tracing::{info_span, trace};
 use tracing_futures::Instrument;
 
 use crate::error::*;
-use crate::handle::*;
 use crate::packet::*;
 use crate::read_req::*;
 #[cfg(feature = "unstable")]
 use crate::write_req::*;
+
+// re-exports
+pub use crate::handle::Handle;
+pub use crate::server_builder::TftpServerBuilder;
 
 pub struct TftpServer<H>
 where
